contracts-stylus: darkpool-core: initial separation of darkpool core contract
