+++ Cargo.toml
@@ -50,6 +50,7 @@ tracing-subscriber = "0.3"
 [profile.release]
 codegen-units = 1        # prefer efficiency to compile time
 panic = "abort"          # use simple panics
+opt-level = "z"
 strip = true             # remove debug info
 lto = true               # link time optimization
 debug = false            # no debug data
@@ -72,3 +73,11 @@ incremental = false # no incremental builds
 # In order to have consistent types in the `arbitrum-client` interfaces,
 # we patch the `contracts-common` dependency to point to the local crate.
 contracts-common = { path = "./contracts-common" }
+
+[patch."https://github.com/renegade-fi/renegade.git"]
+constants = { path = "../renegade/constants" }
+renegade-crypto = { path = "../renegade/renegade-crypto" }
+circuits = { path = "../renegade/circuits" }
+circuit-types = { path = "../renegade/circuit-types" }
+circuit-macros = { path = "../renegade/circuit-macros" }
+arbitrum-client = { path = "../renegade/arbitrum-client" }
